backend/requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asyncpg==0.29.0
bcrypt==5.0.0
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
dnspython==2.8.0
ecdsa==0.19.2
email-validator==2.3.0
fastapi==0.136.0
greenlet==3.4.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.13
iniconfig==2.3.0
joblib==1.3.2
Mako==1.3.11
MarkupSafe==3.0.3
numpy==1.26.3
packaging==26.1
pandas==2.2.0
passlib==1.7.4
pluggy==1.6.0
pyasn1==0.6.3
pycparser==3.0
pydantic==2.13.3
pydantic-settings==2.1.0
pydantic_core==2.46.3
pytest==8.0.0
pytest-asyncio==0.23.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.3.0
python-multipart==0.0.6
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.33.1
resend==2.29.0
rsa==4.9.1
scikit-learn==1.4.0
scipy==1.16.3
sentry-sdk==2.25.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.25
starlette==1.0.0
stripe==7.10.0
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
uvicorn==0.27.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
